Move chat component helpers out of StringUtils
