+++ src/main/java/at/hannibal2/skyhanni/utils/ComponentMatcherUtils.kt
@@ -67,6 +67,9 @@ object ComponentMatcherUtils {
      */
     inline fun <T> Pattern.findStyledMatcher(span: ComponentSpan, consumer: ComponentMatcher.() -> T) =
         styledMatcher(span).let { if (it.find()) consumer(it) else null }
+
+    fun String.applyFormattingFrom(original: ComponentSpan): Component =
+        asComponent { style = original.sampleStyleAtStart() }
 }
 
 /**
